+++ src/main/java/app/candycrisis/search/functions/HeuristicFunction.java
@@ -2,6 +2,6 @@ package app.candycrisis.search.functions;
 
 public interface HeuristicFunction<S> {
 
-    public int estimate(S state);
+    public double estimate(S state);
 
 }
